chore(cypress): Move `clearState` from user utils to shared common utils
